Split out MaterialTypeBuilder to be separate from MaterialBuilder. Common code is now in a new MaterialBuilderUtils.h/cpp, and some is in RPI.Edit's MaterialUtils, as needed according to library dependencies. This is in preparation for the material pipeline abstraction work, see https://github.com/o3de/sig-graphics-audio/blob/main/rfcs/MaterialPipelineAbstraction.md
