+++ Gems/Atom/RPI/Code/Source/RPI.Builders/Model/MaterialAssetBuilderComponent.cpp
@@ -7,6 +7,7 @@
  */
 
 #include <Model/MaterialAssetBuilderComponent.h>
+#include <Material/MaterialBuilderUtils.h>
 
 #include <AzCore/Asset/AssetCommon.h>
 #include <AzCore/Serialization/EditContext.h>
@@ -32,6 +33,7 @@
 
 #include <Atom/RPI.Reflect/Material/MaterialAsset.h>
 #include <AzCore/Settings/SettingsRegistry.h>
+
 
 namespace AZ
 {
@@ -85,7 +87,7 @@ namespace AZ
                 materialTypeSource.m_sourceFileDependencyPath = materialTypePath;
 
                 AssetBuilderSDK::JobDependency jobDependency;
-                jobDependency.m_jobKey = "Atom Material Builder";
+                jobDependency.m_jobKey = "Material Type Builder";
                 jobDependency.m_sourceFile = materialTypeSource;
                 jobDependency.m_platformIdentifier = platformIdentifier;
                 jobDependency.m_productSubIds.push_back(0);
@@ -103,7 +105,7 @@ namespace AZ
 
                 // Note that without the normal job dependencies, materials may not load or reload consistently or in sync with shader and
                 // material type changes without manually monitoring and handling dependency changes.
-                jobDependency.m_type = MaterialUtils::BuildersShouldFinalizeMaterialAssets()
+                jobDependency.m_type = MaterialBuilderUtils::BuildersShouldFinalizeMaterialAssets()
                     ? AssetBuilderSDK::JobDependencyType::Order
                     : AssetBuilderSDK::JobDependencyType::OrderOnce;
 
@@ -119,7 +121,7 @@ namespace AZ
             RPI::MaterialConverterBus::BroadcastResult(conversionInfo, &RPI::MaterialConverterBus::Events::GetFingerprintInfo);
             fingerprintInfo.insert(conversionInfo);
              
-            fingerprintInfo.insert(AZStd::string::format("[BuildersShouldFinalizeMaterialAssets=%d]", MaterialUtils::BuildersShouldFinalizeMaterialAssets()));
+            fingerprintInfo.insert(AZStd::string::format("[BuildersShouldFinalizeMaterialAssets=%d]", MaterialBuilderUtils::BuildersShouldFinalizeMaterialAssets()));
         }
 
         void MaterialAssetBuilderComponent::Reflect(ReflectContext* context)
@@ -232,7 +234,7 @@ namespace AZ
                 }
             }
             
-            MaterialAssetProcessingMode processingMode = MaterialUtils::BuildersShouldFinalizeMaterialAssets() ? MaterialAssetProcessingMode::PreBake : MaterialAssetProcessingMode::DeferredBake;
+            MaterialAssetProcessingMode processingMode = MaterialBuilderUtils::BuildersShouldFinalizeMaterialAssets() ? MaterialAssetProcessingMode::PreBake : MaterialAssetProcessingMode::DeferredBake;
 
             // Build material assets. 
             for (auto& itr : materialSourceDataByUid)
